Finished The Program
Perry, use this to get an idea of what you need to do. Thanks for
uploading what you had to github, This is the corrected version. There
were 5 problems.
No. 1: A Milien Mistake. I have told you to make a new scanner for
different methods. This was what screwed you up there, when its much
easier for you to define it under the CLASS rather than the METHOD. (see
example)

No. 2: Another Milien Mistake. You didn't close the Scanners. Now that
you should be only using 1, you only need to close it once (at the end
of the class) but it's still something to do. Normally not a problem but
you should get into the habit of doing this correctly. This was my fault
becaue I've told you not to worry about it in the past.

No. 3. A Perry Problem. You asked the tourist for their gender, then
used the scanner to assign BOTH Gender and Age, and THEN asked for their
age.

No. 4. A Perry Problem. Don't take the easy way out! I discussed those
tempting little links that offer to make the error report go away at the
click of a button. Don't give into temptation! Use these for reference
only. 

No. 5. A Perry Mistake. Boolean is either TRUE or FALSE. A boolean will
NEVER be anything else. (i.e. not male or female) I corrected this. Part
of this was my fault as I forgot to show you the .equals function for
Strings, which is amazingly useful.